Move run_real and run_virtual to separate module
